backend/requirements.txt
# FastAPI and Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# PDF Processing
PyPDF2>=3.0.0

# AI/ML Dependencies
openai>=1.3.0
numpy>=1.24.0
faiss-cpu>=1.7.4

# Settings Management
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP Requests (for Ollama)
requests>=2.31.0

# Logging (optional but recommended)
python-json-logger>=2.0.7
